Create Endpoints for Sending Password-Resets

+++ routes/api/users.js
@@ -4,11 +4,11 @@ const statusCodes = require("../../config/status_codes");
 const { authenticate, isVerified } = require("../../middleware/security");
 const userMethods = require("../../db_interface/users");
 const jwtDecoding = require("../../helpers/jwt_decoding");
-const {sendVerificationEmail} = require("../../helpers/emailing");
+const { sendVerificationEmail, sendResetEmail } = require("../../helpers/emailing");
 
 // Don't want to log a user in if their email has not been verified
 router.post("/login", isVerified, async (req, res) => {
-  const loginDetails = req.body.user;
+  const loginDetails = req.body;
   console.log("Received login request");
   let valid = false;
   let response = null;
@@ -40,36 +40,22 @@ router.post("/login", isVerified, async (req, res) => {
 router.post("/create", async (req, res) => {
   const userDetails = req.body;
   console.log("Received user-creation request");
-  let userExists = false;
-  let responseJSON = {user: null, errors: []};
+  let responseJSON = { user: null, errors: [] };
   let returnStatus = statusCodes.SUCCESS;
 
   try {
-    let user = await userMethods.getUserByName(userDetails.name);
-    if (user) {
-      userExists = true;
-    } else {
-      user = await userMethods.getUserByEmail(userDetails.email);
-      userExists = user ? true : false;
-    }
-
-    if (userExists) {
+    const newUser = await userMethods.createUser(userDetails);
+    if (newUser.errors.length > 0) {
       returnStatus = statusCodes.INVALID_STATUS;
-      responseJSON.errors.push("User already exists!");
+      responseJSON.errors = newUser.errors;
     } else {
-      const newUser = await userMethods.createUser(userDetails);
-      if (newUser.errors.length > 0) {
-        returnStatus = statusCodes.INVALID_STATUS;
-        responseJSON.errors = newUser.errors;
-      } else {
-        const token = await userMethods.generateEmailVerificationToken(newUser.name);
-        const url = req.protocol + "://" + req.get("Host") + `/api/users/verify/${newUser.name}/${token}`;
-        await sendVerificationEmail(newUser.name, newUser.email, url);
-        responseJSON.user = {
-          name: newUser.name,
-          email: newUser.email
-        };
-      }
+      const token = await userMethods.generateEmailVerificationToken(newUser.name);
+      const url = req.protocol + "://" + req.get("Host") + `/api/users/verify/${newUser.name}/${token}`;
+      sendVerificationEmail(newUser.name, newUser.email, url);
+      responseJSON.user = {
+        name: newUser.name,
+        email: newUser.email,
+      };
     }
   } catch (error) {
     console.error(error.message);
@@ -80,13 +66,13 @@ router.post("/create", async (req, res) => {
   return res.status(returnStatus).json(responseJSON);
 });
 
-router.get("/logout", authenticate, isVerified, async (req, res) => {  
+router.get("/logout", authenticate, isVerified, async (req, res) => {
   const username = jwtDecoding.getUsernameFromToken(jwtDecoding.getTokenFromRequest(req));
   console.log(`Received logout request for ${username}`);
   const errors = await userMethods.deleteJWT(username);
 
-  if(errors.length === 0) {
-    return res.json({ "message": "Successfully Logged Out" });
+  if (errors.length === 0) {
+    return res.json({ message: "Successfully Logged Out" });
   } else {
     return res.status(statusCodes.SERVER_ERROR).json({ errors: errors });
   }
@@ -94,34 +80,36 @@ router.get("/logout", authenticate, isVerified, async (req, res) => {
 
 // This is used when the user's original verification token is lost or expired (i.e. this is how they request a new one)
 // Requires login first.
-router.post("/verify/send/", authenticate, async (req, res) => {
+router.get("/verify/send", authenticate, async (req, res) => {
   const user = res.locals.user;
   const token = await userMethods.generateEmailVerificationToken(user.name);
   const url = req.protocol + "://" + req.get("Host") + `/api/users/verify/${user.name}/${token}`;
-  await sendVerificationEmail(user.name, user.email, url);
-  return res.status(statusCodes.SUCCESS).json({message: "Email verification sent!"});
+  sendVerificationEmail(user.name, user.email, url);
+  return res.status(statusCodes.SUCCESS).json({ message: "Email verification sent!" });
 });
 
 router.get("/verify/:username/:token", async (req, res) => {
   const isValid = await userMethods.verifyEmail(req.params.username, req.params.token);
 
-  if(isValid) {
-    return res.redirect("/login").json({message: "Email verification succeeded!"});
+  if (isValid) {
+    return res.redirect("/login").json({ message: "Email verification succeeded!" });
   } else {
-    return res.status(statusCodes.INVALID_STATUS).json({message: "Email verification failed!"});
+    return res.status(statusCodes.INVALID_STATUS).json({ message: "Email verification failed!" });
   }
 });
 
 router.post("/change-password", authenticate, isVerified, async (req, res) => {
   const password = req.body.password;
-  let responseJSON = {errors: []};
+  const confirmation = req.body.passwordConfirmation;
+  let responseJSON = { errors: [] };
   let returnStatus = statusCodes.SUCCESS;
-  if(userMethods.isValidPassword(password)) {
+
+  if (password === confirmation && userMethods.isValidPassword(password)) {
     try {
       await userMethods.changePassword(res.locals.user, password);
       await userMethods.deleteJWT(res.locals.user.name);
       return res.redirect("../login");
-    } catch(error) {
+    } catch (error) {
       console.error(error.message);
       returnStatus = statusCodes.SERVER_ERROR;
       responseJSON.errors.push("Server Error");
@@ -132,7 +120,57 @@ router.post("/change-password", authenticate, isVerified, async (req, res) => {
     responseJSON.errors.push("Invalid Password");
     return res.status(returnStatus).json(responseJSON);
   }
+});
 
+// This does not require authentication (since the user, by definition, has forgotten their password)
+// Instead, the request must include the reset-password token from the email the user received
+router.post("/reset-password", async (req, res) => {
+  const email = req.body.email;
+  const password = req.body.password;
+  const pwConfirm = req.body.passwordConfirmation;
+  const resetToken = req.body.resetToken;
+
+  let responseJSON = { errors: [] };
+  let returnStatus = statusCodes.SUCCESS;
+
+  if (password === pwConfirm && userMethods.isValidPassword(password)) {
+    try {
+      const user = await userMethods.getUserByEmail(email);
+      if (user && userMethods.verifyForgotPassword(user.name, resetToken)) {
+        await userMethods.changePassword(user, password);
+        await userMethods.deleteJWT(user.name);
+
+        return res.redirect("../login");
+      }
+    } catch (error) {
+      console.error(error.message);
+      returnStatus = statusCodes.SERVER_ERROR;
+      responseJSON.errors.push("Server Error");
+      return res.status(returnStatus).json(responseJSON);
+    }
+  }
+
+  returnStatus = statusCodes.INVALID_STATUS;
+  responseJSON.errors.push("Invalid Request");
+  return res.status(returnStatus).json(responseJSON);
+});
+
+router.post("/request-reset-password", async (req, res) => {
+  const userEmail = req.body.email;
+  try {
+    const user = userMethods.getUserByEmail(userEmail);
+    if (user) {
+      const token = await userMethods.generateForgotPasswordToken(user.name);
+      const url = req.protocol + "://" + req.get("Host") + `/api/users/reset-password/${user.name}/${token}`;
+      sendResetEmail(user.name, user.email, url);
+      return res.status(statusCodes.SUCCESS).json({ message: "Password reset email sent!" });
+    } else {
+      return res.status(statusCodes.INVALID_STATUS).json({ errors: ["User not found"] });
+    }
+  } catch (err) {
+    console.error(err);
+    return res.status(statusCodes.SERVER_ERROR).json({ errors: ["Server Error"] });
+  }
 });
 
 module.exports = router;
